Add test job docker-compose
It is handy to be able to run the tests in continuous integration, add
a job for that. It outputs a percentage of code coverage, as well as
a file that contains a HTML representation of the coverage.

We still need to figure out how to test the error paths.

The tests currently do not run due to package issues, but that is
beyond the scope of this issue.

Closes #9