CLI-rework, shell connecting to kademlia-node
The CLI is now seperated to 2 parts.

The cliserver is started as a goroutine in main.go and is always
listening for a new connection (currently limited to only handling one
connection at the time.)

The cli client is a simple shell and is not part
of the kademlia node, to run it go to src/cli/cliclient and 'run go
cliclient.go' from a terminal.

Partly implements:

- #31 There's no TTL functionality in our code yet.
- #25 Not properly connected to the api, but the argument does reach the
  function that calls the api.
- #24 Connected to the api, but currently returns wether the store was
  successful or not (should return hash if successful according to lab
  spec).

Also partly implements pinging, not properly connected to the api, but
the argument reach the function that calls the api.

Closes #23 entering 'exit' terminates the node by sending the command
through a channel to the loop in main.go to terminate it.

Also implements 'close' which close the connection to the node without
terminating it.  'help' lists all commands the shell currently handles.

The shell is interactive and helps the user through the process of
connecting and how to use commands.

The code should be fairly structured, meaning it's easy to add extra
commands if we need more.