Connect CLI to API
Add a converter from string to KademliaID. Currently no safety checks in
case of bad operation.

Remove prints in the API, as they are not broadcast to the CLI.

Return the key instead of just OK in case of a successful `STORE`.

Format the files in CLI.

Change messages in cliserver and call appropriate API methods. Comment
out forget since it is not implemented.