Test utils/node
Add error checks in hashing.NewKademliaID and hashing.ToKademliaID.

Test the bucket module. The main method that was not covered elsewhere
was that of adding a node to the bucket. There are two main cases to be
aware of, one is when the node was already in the bucket, and the other
when it is seen for the first time. Add tests for both.

Add error checks in node that do not cause the program to crash. Improve
error reporting. Change the node syntax to be node("<ID>","<IP").

Test the node package, particularly the ToString, FromString and
FromStrings methods.

In the routing table there was one method that was untestable, since it
was written in a way as to never hit the final case. Pull out the
hashing.IDLength so that it can be set to 0 and thus hit the final
statement. This is very much a hack and not useful, but it allows
testing all lines of code.

Test the routing table through the TableSynchronizer.